Rename passport-util, install mariadb
