+++ package.json
@@ -31,6 +31,7 @@
     "express-flash": "0.0.2",
     "express-session": "^1.17.1",
     "helmet": "^4.2.0",
+    "mariadb": "^2.5.2",
     "morgan": "^1.10.0",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
